requirements.txt
beautifulsoup4
requests
pandas
sqlalchemy
psycopg2
pytest
python-dotenv